requirements.txt
Flask==1.0.2
flask-mysql
flask-redis
pymysql
redis
jwcrypto
gunicorn
pyyaml
click
